--- a/StringFormatter/Attributes.cs
+++ b/StringFormatter/Attributes.cs
@@ -1,0 +1,3 @@
+﻿using System.Runtime.CompilerServices;
+
+[assembly: InternalsVisibleTo("Test")]