--- a/StringFormatter/GuidFormatting.cs
+++ b/StringFormatter/GuidFormatting.cs
@@ -1,4 +1,5 @@
 using System.Runtime.InteropServices;
+using InlineIL;
 
 namespace System.Text.Formatting
 {
@@ -6,6 +7,8 @@ namespace System.Text.Formatting
     {
         public static unsafe void Format(StringBuffer formatter, Guid value, StringView format)
         {
+            IL.DeclareLocals(false);
+
             if (format.Length > 1)
                 throw new FormatException(string.Format(SR.UnknownFormatSpecifier, format));