--- a/StringFormatter/StringBuffer.cs
+++ b/StringFormatter/StringBuffer.cs
@@ -1,6 +1,7 @@
 ﻿using System.Globalization;
 using System.Reflection;
 using System.Runtime.CompilerServices;
+using InlineIL;
 
 namespace System.Text.Formatting
 {
@@ -569,6 +570,8 @@ namespace System.Text.Formatting
 
         bool AppendSegment<T>(ref char* currRef, char* end, char* dest, ref int prevArgIndex, ref T args) where T : IArgSet
         {
+            IL.DeclareLocals(false);
+
             var curr = currRef;
             var c = '\x0';
             while (curr < end)