--- a/StringFormatter/StringFormatter.csproj
+++ b/StringFormatter/StringFormatter.csproj
@@ -9,6 +9,8 @@
     <NetStandardImplicitPackageVersion>2.0.3</NetStandardImplicitPackageVersion>
   </PropertyGroup>
   <ItemGroup>
+    <PackageReference Include="Fody" Version="4.2.1" PrivateAssets="all" />
+    <PackageReference Include="InlineIL.Fody" Version="1.1.2" PrivateAssets="all" />
     <PackageReference Include="System.Memory" Version="4.5.1" />
     <PackageReference Include="System.Reflection" Version="4.3.0" />
     <PackageReference Include="System.Runtime.CompilerServices.Unsafe" Version="4.5.2" />