--- a/StringFormatter/Unsafe.cs
+++ b/StringFormatter/Unsafe.cs
@@ -1,0 +1,40 @@
+﻿using System.Runtime.CompilerServices;
+using InlineIL;
+using static InlineIL.IL.Emit;
+
+namespace System.Text.Formatting
+{
+    internal static unsafe class Unsafe
+    {
+        [MethodImpl(MethodImplOptions.AggressiveInlining)]
+        public static T Read<T>(void* source)
+        {
+            Ldarg(nameof(source));
+            Ldobj(typeof(T));
+            return IL.Return<T>();
+        }
+
+        [MethodImpl(MethodImplOptions.AggressiveInlining)]
+        public static T As<T>(object o)
+            where T : class
+        {
+            Ldarg(nameof(o));
+            return IL.Return<T>();
+        }
+
+        [MethodImpl(MethodImplOptions.AggressiveInlining)]
+        public static ref TTo As<TFrom, TTo>(ref TFrom source)
+        {
+            Ldarg(nameof(source));
+            return ref IL.ReturnRef<TTo>();
+        }
+
+        [MethodImpl(MethodImplOptions.AggressiveInlining)]
+        public static void* AsPointer<T>(ref T value)
+        {
+            Ldarg(nameof(value));
+            Conv_U();
+            return IL.ReturnPointer();
+        }
+    }
+}