--- a/StringFormatterBench/StringFormatterBench.csproj
+++ b/StringFormatterBench/StringFormatterBench.csproj
@@ -6,9 +6,11 @@
     <AllowUnsafeBlocks>True</AllowUnsafeBlocks>
     <TargetFrameworks>net461;netcoreapp2.1</TargetFrameworks>
   </PropertyGroup>
+
   <ItemGroup>
     <PackageReference Include="BenchmarkDotNet" Version="0.11.3" />
   </ItemGroup>
+
   <ItemGroup>
     <ProjectReference Include="..\StringFormatter\StringFormatter.csproj" />
   </ItemGroup>