--- a/Test/Test.csproj
+++ b/Test/Test.csproj
@@ -5,14 +5,16 @@
     <TargetFrameworks>net461;netcoreapp2.1</TargetFrameworks>
     <LangVersion>latest</LangVersion>
   </PropertyGroup>
+
   <ItemGroup>
-    <PackageReference Include="System.Memory" Version="4.5.1" />
+    <PackageReference Include="System.Runtime.CompilerServices.Unsafe" Version="4.5.2" />
     <PackageReference Include="AutoFixture" Version="4.6.0" />
     <PackageReference Include="Microsoft.NET.Test.Sdk" Version="15.9.0" />
     <PackageReference Include="NFluent" Version="2.4.0" />
     <PackageReference Include="NUnit" Version="3.11.0" />
     <PackageReference Include="NUnit3TestAdapter" Version="3.11.2" />
   </ItemGroup>
+
   <ItemGroup>
     <ProjectReference Include="..\StringFormatter\StringFormatter.csproj" />
   </ItemGroup>