This PR clears the "init locals" flag in all methods that use `stackalloc` which causes the JIT to emit code which zeroes the memory unnecessarily. Some of these stack-allocated regions are quite large.

`StringFormatterBench` Before:

```
BenchmarkDotNet=v0.11.3, OS=Windows 10.0.17763.439 (1809/October2018Update/Redstone5)
Intel Core i7-7700 CPU 3.60GHz (Kaby Lake), 1 CPU, 8 logical and 4 physical cores
  [Host]       : .NET Framework 4.7.2 (CLR 4.0.30319.42000), 64bit RyuJIT-v4.8.3761.0
  LegacyJitX64 : .NET Framework 4.7.2 (CLR 4.0.30319.42000), 64bit LegacyJIT/clrjit-v4.8.3761.0;compatjit-v4.8.3761.0
  RyuJitX64    : .NET Framework 4.7.2 (CLR 4.0.30319.42000), 64bit RyuJIT-v4.8.3761.0

Platform=X64  Runtime=Clr

       Method |          Job |       Jit |       Mean |     Error |   StdDev |        Min |        Max | Ratio | Gen 0/1k Op | Gen 1/1k Op | Gen 2/1k Op | Allocated Memory/Op |
------------- |------------- |---------- |-----------:|----------:|---------:|-----------:|-----------:|------:|------------:|------------:|------------:|--------------------:|
     Baseline | LegacyJitX64 | LegacyJit | 1,063.9 ns |  9.192 ns | 8.598 ns | 1,049.0 ns | 1,075.6 ns |  1.00 |      0.1335 |           - |           - |               562 B |
 StringBuffer | LegacyJitX64 | LegacyJit |   724.9 ns |  2.256 ns | 2.000 ns |   722.5 ns |   728.6 ns |  0.68 |      0.0324 |           - |           - |               136 B |
              |              |           |            |           |          |            |            |       |             |             |             |                     |
     Baseline |    RyuJitX64 |    RyuJit | 1,068.7 ns | 10.795 ns | 9.569 ns | 1,056.2 ns | 1,088.3 ns |  1.00 |      0.1335 |           - |           - |               562 B |
 StringBuffer |    RyuJitX64 |    RyuJit |   728.2 ns |  4.138 ns | 3.871 ns |   724.1 ns |   736.4 ns |  0.68 |      0.0324 |           - |           - |               136 B |
```

After:

```
       Method |          Job |       Jit |       Mean |     Error |    StdDev |        Min |        Max | Ratio | Gen 0/1k Op | Gen 1/1k Op | Gen 2/1k Op | Allocated Memory/Op |
------------- |------------- |---------- |-----------:|----------:|----------:|-----------:|-----------:|------:|------------:|------------:|------------:|--------------------:|
     Baseline | LegacyJitX64 | LegacyJit | 1,065.8 ns |  7.324 ns |  6.116 ns | 1,057.0 ns | 1,077.9 ns |  1.00 |      0.1335 |           - |           - |               562 B |
 StringBuffer | LegacyJitX64 | LegacyJit |   653.1 ns |  7.332 ns |  6.858 ns |   646.5 ns |   666.6 ns |  0.61 |      0.0324 |           - |           - |               136 B |
              |              |           |            |           |           |            |            |       |             |             |             |                     |
     Baseline |    RyuJitX64 |    RyuJit | 1,078.3 ns | 12.035 ns | 10.668 ns | 1,067.6 ns | 1,104.6 ns |  1.00 |      0.1335 |           - |           - |               562 B |
 StringBuffer |    RyuJitX64 |    RyuJit |   609.0 ns |  2.745 ns |  2.568 ns |   605.9 ns |   614.7 ns |  0.56 |      0.0324 |           - |           - |               136 B |
```

I used InlineIL.Fody for now to achieve this. This could be replaced with [this](https://github.com/dotnet/csharplang/blob/master/proposals/skip-localsinit.md) compiler feature in the future.

I also cleaned up some stuff and removed some dependencies.